Move the code into a package subfolder
