+++ src/hashbeaf/hashbeaf.py
@@ -2,8 +2,8 @@ import argparse
 import logging
 from typing import List, Tuple
 
-import utils
-import words
+from .utils import get_hash, ishex, run_command
+from .words import WORDS
 
 logger = logging.getLogger(__name__)
 
@@ -27,24 +27,24 @@ def _commit_data_modify_increment(
 def hashbeaf_main(words: List[str], max_minutes_in_future: int) -> None:
     words = [word.lower() for word in words]
     for word in words:
-        if not utils.ishex(word):
+        if not ishex(word):
             raise RuntimeError(
                 f"User input '{word}' contains non-hexadecimal characters, aborting."
             )
 
-    commit_data_original = utils.run_command("git cat-file commit HEAD")
+    commit_data_original = run_command("git cat-file commit HEAD")
     for commit_increment in range(max_minutes_in_future * 60):
         for author_increment in range(commit_increment + 1):
             (commit_data, author_time_full, commit_time_full) = _commit_data_modify_increment(
                 commit_data_original, author_increment, commit_increment
             )
-            new_hash = utils.get_hash(commit_data)
+            new_hash = get_hash(commit_data)
             for word in words:
                 if new_hash.startswith(word):
                     if commit_increment == author_increment == 0:
                         logger.info("The current hash is already nice, nothing to do")
                         return
-                    result = utils.run_command(
+                    result = run_command(
                         f"GIT_COMMITTER_DATE='{commit_time_full}' "
                         f"git commit --amend -C HEAD --date='{author_time_full}'"
                     )
@@ -54,10 +54,14 @@ def hashbeaf_main(words: List[str], max_minutes_in_future: int) -> None:
     logger.info("Did not find a nice hash, retry with a shorter word or more minutes in the future")
 
 
-if __name__ == "__main__":
+def cli_main() -> None:
     logging.basicConfig()
     logging.getLogger().setLevel(logging.INFO)
     parser = argparse.ArgumentParser("Change the last commit to a commit hash with words you like")
-    parser.add_argument("words", type=str, nargs="?", default=words.WORDS)
+    parser.add_argument("words", type=str, nargs="?", default=WORDS)
     parser.add_argument("--max_minutes_in_future", type=int, default=15)
     hashbeaf_main(**vars(parser.parse_args()))
+
+
+if __name__ == "__main__":
+    cli_main()
