+++ src/hashbeaf/words.py
